Update repo from tezos-k8s (#1)
* Exit config-init script on error (#531)

* Exit config-init script on error

* Update chart tests

* Set fitness to 1 instead of -1 at activation (#534)

A bug was uncovered during mumbai activation, where nodes were unable to
activate and were crashing.

The [fix](https://gitlab.com/tezos/tezos/-/merge_requests/7462/diffs)
broke dailynet because it is no longer possible to activate with a
fitness of -1 (which is how we have been activating chains since the
beginning of tezos-k8s).

Replacing this value to 1 fixes the issue for dailynet and also does not
appear to be causing any issue with v15.1.

* tezos-proto-cruncher chart (#532)

* tezos-proto-cruncher chart

This chart deploys a daemonset in your cluster. the daemonset runs a
parallel search for a vanity protocol hash.

It takes s3 bucket credentials. Protocol is downloaded from bucket then
results are written to the bucket.

It does not terminate: it keeps running until the cluster is torn down.

* better variable names

* Update charts/tezos-proto-cruncher/scripts/proto-downloader.py

Co-authored-by: Aryeh Harris <harryttd@users.noreply.github.com>

* address review comments

* lint python scripts with black

* catch exception when s3 upload fails

* fix manual method

* fix manual method, better way per Aryeh

* fail fast when env vars not set in proto downloader

Co-authored-by: Aryeh Harris <harryttd@users.noreply.github.com>

Co-authored-by: Aryeh Harris <harryttd@users.noreply.github.com>
Co-authored-by: Nicolas Ochem <nicolas.ochem@gmail.com>